+++ src/components/DataView/DataViewActionsButtonRow/DataViewActionsButtonRowTypes.ts
@@ -1,0 +1,9 @@
+import { DataViewAction, DataViewAdditionalAction, DataViewControlViewOption } from "../DataViewTypes";
+import { MosaicObject } from "../../../types";
+
+export interface DataViewActionsButtonRowProps {
+	primaryActions: DataViewAction[];
+	additionalActions: DataViewAdditionalAction[];
+	originalRowData: MosaicObject;
+	activeDisplay?: DataViewControlViewOption;
+}
