+++ src/components/DataView/DataViewActionsButtonRow/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./DataViewActionsButtonRow";
+export * from "./DataViewActionsButtonRow";
+export * from "./DataViewActionsButtonRowTypes";
